+++ app/src/main/java/com/tari/android/wallet/di/ApplicationComponent.kt
@@ -47,6 +47,7 @@ import com.tari.android.wallet.ui.fragment.contact_book.details.ContactDetailsVi
 import com.tari.android.wallet.ui.fragment.contact_book.link.ContactLinkViewModel
 import com.tari.android.wallet.ui.fragment.contact_book.root.ContactBookViewModel
 import com.tari.android.wallet.ui.fragment.contact_book.root.ShareViewModel
+import com.tari.android.wallet.ui.fragment.contact_book.transactionHistory.TransactionHistoryViewModel
 import com.tari.android.wallet.ui.fragment.home.HomeActivity
 import com.tari.android.wallet.ui.fragment.home.HomeViewModel
 import com.tari.android.wallet.ui.fragment.onboarding.activity.OnboardingFlowActivity
@@ -90,6 +91,7 @@ import com.tari.android.wallet.ui.fragment.settings.torBridges.TorBridgesSelecti
 import com.tari.android.wallet.ui.fragment.settings.torBridges.customBridges.CustomTorBridgesViewModel
 import com.tari.android.wallet.ui.fragment.settings.userAutorization.BiometricAuthenticationViewModel
 import com.tari.android.wallet.ui.fragment.splash.SplashActivity
+import com.tari.android.wallet.ui.fragment.tx.TransactionRepository
 import com.tari.android.wallet.ui.fragment.tx.TxListViewModel
 import com.tari.android.wallet.ui.fragment.tx.details.TxDetailsViewModel
 import com.tari.android.wallet.ui.fragment.tx.details.gif.GIFViewModel
@@ -181,6 +183,8 @@ interface ApplicationComponent {
     fun inject(viewModel: ContactSelectionViewModel)
     fun inject(viewModel: TxDetailsViewModel)
     fun inject(viewModel: ShareViewModel)
+    fun inject(viewModel: TransactionRepository)
+    fun inject(viewModel: TransactionHistoryViewModel)
 
     fun getClipboardManager(): ClipboardManager
 }
