+++ app/src/main/java/com/tari/android/wallet/ui/fragment/home/navigation/Navigation.kt
@@ -111,6 +111,8 @@ sealed class Navigation {
 
         class ToLinkContact(val contact: ContactDto) : ContactBookNavigation()
 
+        class ToContactTransactionHistory(val contact: ContactDto) : ContactBookNavigation()
+
         class BackToContactBook : ContactBookNavigation()
     }
 }
