+++ app/src/main/res/drawable/vector_empty_wallet.xml
@@ -1,0 +1,17 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="151dp"
+    android:height="167dp"
+    android:viewportWidth="151"
+    android:viewportHeight="167">
+  <path
+      android:pathData="M9.4,64.11C6.43,58.96 8.19,52.37 13.35,49.4L96.4,1.44C101.55,-1.53 108.14,0.24 111.11,5.39C114.09,10.54 112.32,17.13 107.17,20.1L95.87,26.63C99.02,27.09 101.94,28.93 103.66,31.9C106.09,36.11 105.36,41.27 102.21,44.64L132.54,27.13C137.69,24.16 144.28,25.92 147.25,31.07C150.23,36.23 148.46,42.81 143.31,45.79L120.68,58.85C124.63,58.61 128.57,60.56 130.68,64.22C133.65,69.37 131.89,75.96 126.74,78.93L115.5,85.42C119.72,84.82 124.08,86.78 126.34,90.7C128.57,94.55 128.14,99.2 125.64,102.54L134.34,97.52C139.49,94.55 146.08,96.31 149.05,101.47C152.03,106.62 150.26,113.21 145.11,116.18L62.06,164.13C56.9,167.11 50.32,165.34 47.34,160.19C45.12,156.34 45.54,151.7 48.04,148.35L19.78,164.67C14.63,167.64 8.04,165.88 5.07,160.72C2.09,155.57 3.86,148.98 9.01,146.01L20.25,139.52C16.02,140.12 11.67,138.16 9.4,134.24C6.43,129.09 8.19,122.5 13.35,119.52L35.97,106.46C32.02,106.7 28.09,104.75 25.98,101.09C23.55,96.89 24.28,91.73 27.43,88.35L16.66,94.57C11.51,97.54 4.92,95.78 1.94,90.63C-1.03,85.48 0.74,78.89 5.89,75.91L17.19,69.39C14.04,68.93 11.12,67.08 9.4,64.11Z"
+      android:strokeAlpha="0.3"
+      android:fillColor="#9330FF"
+      android:fillType="evenOdd"
+      android:fillAlpha="0.3"/>
+  <path
+      android:pathData="M25.33,56.41V125.02C25.33,126.49 26.53,127.68 28,127.68H124C125.47,127.68 126.67,126.49 126.67,125.02V106.41H86.67C79.3,106.41 73.33,100.46 73.33,93.11C73.33,85.77 79.3,79.82 86.67,79.82H126.67V61.34C126.67,59.9 125.51,58.72 124.07,58.69L118.5,58.54H32.26C29.69,58.54 27.3,57.75 25.33,56.4L25.33,56.41ZM20.05,47.4C20.02,47.04 20,46.68 20,46.31V45.25C20,37.91 25.97,31.95 33.33,31.95H49.33C50.81,31.95 52,33.14 52,34.61C52,36.08 50.81,37.27 49.33,37.27H33.33C28.91,37.27 25.33,40.84 25.33,45.25V46.31C25.33,50.13 28.44,53.23 32.27,53.23H116V37.27H102.67C101.19,37.27 100,36.08 100,34.61C100,33.14 101.19,31.95 102.67,31.95H118.67C120.14,31.95 121.33,33.14 121.33,34.61V53.3L124.2,53.37C128.54,53.48 132,57.02 132,61.34V125.02C132,129.43 128.42,133 124,133H28C23.58,133 20,129.43 20,125.02V47.91C20,47.74 20.02,47.57 20.05,47.4L20.05,47.4ZM126.67,101.09V85.14H86.67C82.25,85.14 78.67,88.71 78.67,93.11C78.67,97.52 82.25,101.09 86.67,101.09L126.67,101.09ZM78.67,34.61C78.67,36.08 77.47,37.27 76,37.27C74.53,37.27 73.33,36.08 73.33,34.61V18.66C73.33,17.19 74.53,16 76,16C77.47,16 78.67,17.19 78.67,18.66V34.61ZM64.11,33.12C64.73,34.45 64.15,36.03 62.82,36.65C61.48,37.27 59.9,36.69 59.27,35.36L55.89,28.13C55.27,26.8 55.85,25.22 57.18,24.6C58.52,23.98 60.1,24.56 60.73,25.89L64.11,33.12ZM92.73,35.36C92.1,36.69 90.52,37.27 89.18,36.65C87.85,36.03 87.27,34.45 87.89,33.12L91.27,25.89C91.9,24.56 93.48,23.98 94.82,24.6C96.15,25.22 96.73,26.8 96.11,28.13L92.73,35.36ZM92,98.43C89.05,98.43 86.67,96.05 86.67,93.11C86.67,90.18 89.05,87.79 92,87.79C94.95,87.79 97.33,90.18 97.33,93.11C97.33,96.05 94.95,98.43 92,98.43Z"
+      android:strokeAlpha="0.7"
+      android:fillColor="#000000"
+      android:fillAlpha="0.7"/>
+</vector>
