+++ app/src/main/res/layout/fragment_contact_transaction_history.xml
@@ -1,0 +1,98 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<com.tari.android.wallet.ui.component.tari.TariPrimaryBackground xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/root_view"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <androidx.appcompat.widget.LinearLayoutCompat
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical">
+
+        <com.tari.android.wallet.ui.component.tari.toolbar.TariToolbar
+            android:id="@+id/toolbar"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:text="@string/contact_details_transaction_history" />
+
+        <com.tari.android.wallet.ui.component.tari.TariSecondaryBackground
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <androidx.appcompat.widget.LinearLayoutCompat
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:orientation="vertical">
+
+                <com.tari.android.wallet.ui.component.tari.TariTextView
+                    android:id="@+id/description_view"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_marginHorizontal="@dimen/common_dialog_horizontal_margin"
+                    android:layout_marginVertical="20dp"
+                    android:layout_marginBottom="20dp"
+                    android:gravity="center_horizontal"
+                    android:text="@string/contact_details_transaction_history_description"
+                    android:textColor="?attr/palette_text_body"
+                    app:customFont="medium" />
+
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/list"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content" />
+
+                <androidx.appcompat.widget.LinearLayoutCompat
+                    android:id="@+id/empty_state"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:gravity="center"
+                    android:orientation="vertical"
+                    android:visibility="gone"
+                    tools:visibility="visible">
+
+                    <androidx.appcompat.widget.AppCompatImageView
+                        android:layout_width="150dp"
+                        android:layout_height="166dp"
+                        android:layout_marginBottom="40dp"
+                        android:src="@drawable/vector_empty_wallet" />
+
+                    <com.tari.android.wallet.ui.component.tari.TariTextView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginBottom="20dp"
+                        android:gravity="center"
+                        android:text="@string/contact_details_transaction_history_empty_state_title"
+                        android:textColor="?attr/palette_text_heading"
+                        android:textSize="18sp"
+                        app:customFont="regular" />
+
+                    <com.tari.android.wallet.ui.component.tari.TariTextView
+                        android:id="@+id/empty_state_description"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginHorizontal="35dp"
+                        android:layout_marginBottom="20dp"
+                        android:gravity="center"
+                        android:text="@string/contact_details_transaction_history_empty_state_description"
+                        android:textColor="?attr/palette_text_body"
+                        android:lineSpacingExtra="8dp"
+                        app:customFont="medium" />
+
+                    <com.tari.android.wallet.ui.component.tari.TariSecondaryButton
+                        android:id="@+id/send_tari_button"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        app:title="@string/send_tari_title" />
+
+                </androidx.appcompat.widget.LinearLayoutCompat>
+
+            </androidx.appcompat.widget.LinearLayoutCompat>
+
+        </com.tari.android.wallet.ui.component.tari.TariSecondaryBackground>
+
+    </androidx.appcompat.widget.LinearLayoutCompat>
+
+</com.tari.android.wallet.ui.component.tari.TariPrimaryBackground>
